refactor(cli): give the roadmap table its own reader
C6's framing clause put `product-lint.js` at 772 lines, over the 700-line hard
limit in `.claude/rules/coding-standards.md`. Nothing counts lines in CI, which
is why `docs/plans/check-product-deferred.md` §4 exists as a standing
constraint — this is the second time it has bound, and the section records both.

`scripts/lib/product-roadmap.js` takes the roadmap table's reader:
`parseRoadmapRows`, `rowCells`, `checkRoadmapFraming`, `roadmapSection` and the
table's own constants. The cut is by on-disk format rather than by
markdown-versus-product-state (which is what `product-markdown.js` already
splits on), so the six-column table has one owner and the two rules about the
table's *shape* rather than about what its rows say — C4's arity check and C6's
framing clause — travel with it. `product-lint.js` comes out at 635 and keeps
every rule that reads the parsed state.

No behaviour change, and the suite is the proof: `check-product.test.js`
requires only `validateProduct`, so a green run across the move shows the split
changed nothing. `scripts/check-product.js` imports `parseRoadmapRows` from the
new module directly rather than through a re-export, per the no-barrel rule, and
`ROW_STATUSES` moves with it for the same reason.
